api,config,sign: Organize bootcfg packages together
