+++ bootcfg/api/cloud_test.go
