+++ bootcfg/api/ipxe_test.go
