+++ bootcfg/api/pixiecore_test.go
