+++ bootcfg/api/serialize_test.go
