+++ bootcfg/sign/writer_test.go
